# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ReLink AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface AnthropicConversationInput {
  title: string;
}

export interface AnthropicMessageInput {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export type RelationStatus = typeof RelationStatus[keyof typeof RelationStatus];


export const RelationStatus = {
  active: 'active',
  paused: 'paused',
  resolved: 'resolved',
} as const;

export interface Relation {
  id: number;
  name: string;
  participantMe: string;
  participantOther: string;
  status: RelationStatus;
  messageCount: number;
  lastMessageAt?: string | null;
  memoryBuiltAt?: string | null;
  createdAt: string;
}

export type RelationDetailStatus = typeof RelationDetailStatus[keyof typeof RelationDetailStatus];


export const RelationDetailStatus = {
  active: 'active',
  paused: 'paused',
  resolved: 'resolved',
} as const;

export interface RelationDetail {
  id: number;
  name: string;
  participantMe: string;
  participantOther: string;
  status: RelationDetailStatus;
  messageCount: number;
  lastMessageAt?: string | null;
  memoryBuiltAt?: string | null;
  createdAt: string;
  currentPhase?: string | null;
  summary?: string | null;
}

export interface CreateRelationInput {
  name: string;
  participantMe: string;
  participantOther: string;
}

export interface WhatsappImportInput {
  /** Raw content of the WhatsApp export file */
  content: string;
  filename?: string;
}

export interface PasteImportInput {
  text: string;
  participantMe: string;
  participantOther: string;
}

export interface ScreenshotImportInput {
  /** Base64 encoded image */
  imageBase64: string;
  mimeType?: string;
}

export interface ExtractedMessage {
  tempId: string;
  sender: string;
  content: string;
  timestamp?: string | null;
  isMe: boolean;
}

export interface ScreenshotExtractResult {
  extractedMessages: ExtractedMessage[];
  confidence: number;
  needsReview: boolean;
}

export interface ScreenshotConfirmInput {
  messages: ExtractedMessage[];
  defaultDate?: string;
}

export interface ManualMessageInput {
  sender: string;
  content: string;
  isMe: boolean;
  sentAt: string;
}

export type WhatsappMessageImportSource = typeof WhatsappMessageImportSource[keyof typeof WhatsappMessageImportSource];


export const WhatsappMessageImportSource = {
  whatsapp_file: 'whatsapp_file',
  paste: 'paste',
  screenshot: 'screenshot',
  manual: 'manual',
} as const;

export interface WhatsappMessage {
  id: number;
  relationId: number;
  sender: string;
  content: string;
  isMe: boolean;
  sentAt: string;
  importSource: WhatsappMessageImportSource;
}

export type ImportResultDateRange = {
  from: string | null;
  to: string | null;
};

export interface ImportResult {
  imported: number;
  duplicates: number;
  totalMessages: number;
  dateRange: ImportResultDateRange;
}

export interface PaginatedMessages {
  messages: WhatsappMessage[];
  nextCursor?: string | null;
  total: number;
}

export type MessageStatsByParticipant = {[key: string]: number};

export type MessageStatsDateRange = {
  from: string | null;
  to: string | null;
};

export interface MessageStats {
  total: number;
  byParticipant: MessageStatsByParticipant;
  dateRange: MessageStatsDateRange;
  avgResponseTimeMinutes?: number | null;
  daysWithMessages: number;
}

export type RelationalMemoryCommunicationTrends = {
  whoInitiates?: string | null;
  responseBalance?: string | null;
  overallTone?: string | null;
} | null;

export interface RelationalMemory {
  relationId: number;
  globalSummary?: string | null;
  currentPhase?: string | null;
  recurringTopics: string[];
  expressedLimits: string[];
  openQuestions: string[];
  importantEvents: string[];
  communicationTrends?: RelationalMemoryCommunicationTrends;
  builtAt?: string | null;
  isBuilding: boolean;
}

export interface RelationPhase {
  id: number;
  label: string;
  description: string;
  startDate?: string | null;
  endDate?: string | null;
  isCurrentPhase: boolean;
}

export interface DynamicReport {
  whoInitiates?: string | null;
  whoFollowsUp?: string | null;
  avgResponseTime?: string | null;
  messageFrequency?: string | null;
  recentChanges?: string | null;
  recurringConflicts: string[];
  silences: string[];
  expressedLimits: string[];
  unansweredQuestions: string[];
  observableFacts: string[];
  trends: string[];
  hypotheses: string[];
  unknowns: string[];
}

export interface AgentSession {
  id: number;
  relationId: number;
  title: string;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export type AgentMessageRole = typeof AgentMessageRole[keyof typeof AgentMessageRole];


export const AgentMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface AgentMessage {
  id: number;
  sessionId: number;
  role: AgentMessageRole;
  content: string;
  contextUsed?: string[] | null;
  createdAt: string;
}

export interface AgentSessionWithMessages {
  id: number;
  relationId: number;
  title: string;
  createdAt: string;
  messages: AgentMessage[];
}

export interface CreateAgentSessionInput {
  title: string;
  initialMessage?: string | null;
}

export interface AgentChatInput {
  message: string;
  selectedMessageIds?: number[] | null;
}

export interface AnalyzeMessageInput {
  messageId?: number | null;
  messageContent?: string | null;
  additionalContext?: string | null;
}

export interface SuggestionsInput {
  messageId?: number | null;
  messageContent?: string | null;
  context?: string | null;
}

export type ResponseSuggestionStyle = typeof ResponseSuggestionStyle[keyof typeof ResponseSuggestionStyle];


export const ResponseSuggestionStyle = {
  natural: 'natural',
  calm: 'calm',
  firm: 'firm',
} as const;

export interface ResponseSuggestion {
  id: string;
  text: string;
  style: ResponseSuggestionStyle;
  intention: string;
  tone: string;
  explanation: string;
}

export interface SuggestionsResult {
  suggestions: ResponseSuggestion[];
  analysis?: string | null;
}

export type TransformSuggestionInputTransformation = typeof TransformSuggestionInputTransformation[keyof typeof TransformSuggestionInputTransformation];


export const TransformSuggestionInputTransformation = {
  shorter: 'shorter',
  more_natural: 'more_natural',
  softer: 'softer',
  firmer: 'firmer',
  no_blame: 'no_blame',
  single_question: 'single_question',
  detached: 'detached',
  dignified: 'dignified',
} as const;

export interface TransformSuggestionInput {
  originalText: string;
  transformation: TransformSuggestionInputTransformation;
  relationId: number;
}

export type ListMessagesParams = {
cursor?: string;
limit?: number;
search?: string;
date?: string;
};