# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ReLink AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /relations/{relationId}/agent/sessions:
    get:
      operationId: listAgentSessions
      tags: [agent]
      summary: List agent chat sessions for a relation
      parameters:
        - name: relationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentSession"
    post:
      operationId: createAgentSession
      tags: [agent]
      summary: Start a new agent session
      parameters:
        - name: relationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentSessionInput"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSession"
  /relations/{relationId}/agent/sessions/{sessionId}:
    get:
      operationId: getAgentSession
      tags: [agent]
      summary: Get agent session with messages
      parameters:
        - name: relationId
          in: path
          required: true
          schema:
            type: integer
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSessionWithMessages"
  /relations/{relationId}/agent/sessions/{sessionId}/chat:
    post:
      operationId: chatWithAgent
      tags: [agent]
      summary: Send a message to ReLink AI (SSE stream)
      parameters:
        - name: relationId
          in: path
          required: true
          schema:
            type: integer
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentChatInput"
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream: {}

  # ── Analysis ───────────────────────────────────────────────────────────────
  /relations/{relationId}/analysis/message:
    post:
      operationId: analyzeMessage
      tags: [analysis]
      summary: Analyze a specific message
      parameters:
        - name: relationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeMessageInput"
      responses:
        "200":
          description: Message analysis (SSE)
          content:
            text/event-stream: {}
  /relations/{relationId}/analysis/suggestions:
    post:
      operationId: getSuggestions
      tags: [analysis]
      summary: Get response suggestions
      parameters:
        - name: relationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestionsInput"
      responses:
        "200":
          description: Response suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestionsResult"
  /relations/{relationId}/analysis/suggestions/transform:
    post:
      operationId: transformSuggestion
      tags: [analysis]
      summary: Transform a response suggestion
      parameters:
        - name: relationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransformSuggestionInput"
      responses:
        "200":
          description: Transformed suggestion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseSuggestion"
components:
  schemas:
...
    AgentSession:
      type: object
      properties:
        id:
          type: integer
        relationId:
          type: integer
        title:
          type: string
        messageCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, relationId, title, messageCount, createdAt, updatedAt]
    AgentSessionWithMessages:
      type: object
      properties:
        id:
          type: integer
        relationId:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AgentMessage"
      required: [id, relationId, title, createdAt, messages]
    AgentMessage:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        contextUsed:
          type: array
          items:
            type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, sessionId, role, content, createdAt]
    CreateAgentSessionInput:
      type: object
      properties:
        title:
          type: string
        initialMessage:
          type: string
          nullable: true
      required: [title]
    AgentChatInput:
      type: object
      properties:
        message:
          type: string
        selectedMessageIds:
          type: array
          items:
            type: integer
          nullable: true
      required: [message]

    # ── Analysis ──────────────────────────────────────────────────────────────
    AnalyzeMessageInput:
      type: object
      properties:
        messageId:
          type: integer
          nullable: true
        messageContent:
          type: string
          nullable: true
        additionalContext:
          type: string
          nullable: true
      required: []
    SuggestionsInput:
      type: object
      properties:
        messageId:
          type: integer
          nullable: true
        messageContent:
          type: string
          nullable: true
        context:
          type: string
          nullable: true
      required: []
    ResponseSuggestion:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        style:
          type: string
          enum: [natural, calm, firm]
        intention:
          type: string
        tone:
          type: string
        explanation:
          type: string
      required: [id, text, style, intention, tone, explanation]
    SuggestionsResult:
      type: object
      properties:
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/ResponseSuggestion"
        analysis:
          type: string
          nullable: true
      required: [suggestions]
    TransformSuggestionInput:
      type: object
      properties:
        originalText:
          type: string
        transformation:
          type: string
          enum: [shorter, more_natural, softer, firmer, no_blame, single_question, detached, dignified]
        relationId:
          type: integer
      required: [originalText, transformation, relationId]